Initial commit - completed project
This project was created as an assignment for Jetbrains Academy's Python programming track.
The objective was to create a tic tac toe game with support for 2-player play, and player vs computer play.
The computer player has three levels of difficult, the hardest of which uses minimax to find the best possible move.